requirements.txt
Pillow~=9.1.0
PyAutoGUI~=0.9.53
selenium~=4.1.3